+++ src/Compiler/Service/ServiceParsedInputOps.fs
@@ -1633,7 +1633,7 @@ module ParsedInput =
                             None
 
                     // module Namespace.Top
-                    // module Neste|
+                    // module Nested
                     | SynModuleDecl.NestedModule(moduleInfo = SynComponentInfo(longId = [ ident ])) when rangeContainsPos ident.idRange pos ->
                         Some CompletionContext.Invalid