+++ src/Compiler/TypedTree/TcGlobals.fs
@@ -1055,7 +1055,7 @@ type TcGlobals(
             | true, builder -> builder tinst
             | _ -> TType_app (tcref, tinst, v_knownWithoutNull)
 
-  // Adding an unnecessary "let" instead of inlining into a muiti-line pipelined compute-once "member val" that is too complex for @dsyme
+  // Adding an unnecessary "let" instead of inlining into a multi-line pipelined compute-once "member val" that is too complex for @dsyme
   let v_attribs_Unsupported = [
         tryFindSysAttrib "System.Runtime.CompilerServices.ModuleInitializerAttribute"
         tryFindSysAttrib "System.Runtime.CompilerServices.CallerArgumentExpressionAttribute"