+++ tests/FSharp.Compiler.ComponentTests/Diagnostics/Records.fs
@@ -72,14 +72,14 @@ let ``Warning not emitted for generated record updates within a nested copy-and-
     Fsx """
 type AnotherNestedRecTy = { A: int; B: int }
 
-type NestdRecTy = { C: {| c: AnotherNestedRecTy |} }
+type NestedRecTy = { C: {| c: AnotherNestedRecTy |} }
 
-type RecTy = { D: NestdRecTy; I: int }
+type RecTy = { D: NestedRecTy; I: int }
 
-//                       vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv
-let t1 (x: NestdRecTy) = { x with C.c = Unchecked.defaultof<_> }
+//                        vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv
+let t1 (x: NestedRecTy) = { x with C.c = Unchecked.defaultof<_> }
 
-// Do not report for the nested NestdRecTy update
+// Do not report for the nested NestedRecTy update
 let t2 (x: RecTy) (a: AnotherNestedRecTy) = { x with D.C.c = { a with A = 3 } }
 
 //                                                           vvvvvvvvvvvvvvvvvvvvvvv
@@ -90,7 +90,7 @@ let t3 (x: RecTy) (a: AnotherNestedRecTy) = { x with D.C.c = { a with A = 3; B =
     |> typecheck
     |> shouldFail
     |> withDiagnostics [
-        (Warning 3560, Line 9, Col 26, Line 9, Col 65, "This copy-and-update record expression changes all fields of record type 'Test.NestdRecTy'. Consider using the record construction syntax instead.")
+        (Warning 3560, Line 9, Col 27, Line 9, Col 66, "This copy-and-update record expression changes all fields of record type 'Test.NestedRecTy'. Consider using the record construction syntax instead.")
         (Warning 3560, Line 15, Col 62, Line 15, Col 85, "This copy-and-update record expression changes all fields of record type 'Test.AnotherNestedRecTy'. Consider using the record construction syntax instead.")
     ]