+++ tests/FSharp.Compiler.ComponentTests/Interop/DeeplyNestedCSharpClasses.fs
@@ -33,7 +33,7 @@ namespace MyNamespace
 
         let fsharpSource =
             """
-let loss2 = MyNamespace.OoterClass.InnerClass.MoreInnerClass.somefunction()   //Note the miss-typed functionname we expect a good error message
+let loss2 = MyNamespace.OoterClass.InnerClass.MoreInnerClass.somefunction()   //Note the mistyped functionname we expect a good error message
 """
         FSharp fsharpSource
         |> asExe
@@ -47,7 +47,7 @@ let loss2 = MyNamespace.OoterClass.InnerClass.MoreInnerClass.somefunction()   //
 
         let fsharpSource =
             """
-let loss2 = MyNamespace.OuterClass.InnerClass_.MoreInnerClass.somefunction()   //Note the miss-typed InnerClass name we expect a good error message
+let loss2 = MyNamespace.OuterClass.InnerClass_.MoreInnerClass.somefunction()   //Note the mistyped InnerClass name we expect a good error message
 """
         FSharp fsharpSource
         |> asExe
@@ -61,7 +61,7 @@ let loss2 = MyNamespace.OuterClass.InnerClass_.MoreInnerClass.somefunction()   /
 
         let fsharpSource =
             """
-let loss2 = MyNamespace.OuterClass.InnerClass.MoareInnerClass.somefunction()   //Note the miss-typed MoreInnerClass we expect a good error message
+let loss2 = MyNamespace.OuterClass.InnerClass.MoareInnerClass.somefunction()   //Note the mistyped MoreInnerClass we expect a good error message
 """
         FSharp fsharpSource
         |> asExe
@@ -75,7 +75,7 @@ let loss2 = MyNamespace.OuterClass.InnerClass.MoareInnerClass.somefunction()   /
 
         let fsharpSource =
             """
-let loss2 = MyNamespace.OuterClass.InnerClass.MoreInnerClass.somefunctoion()   //Note the miss-typed somefunction we expect a good error message
+let loss2 = MyNamespace.OuterClass.InnerClass.MoreInnerClass.somefunctoion()   //Note the mistyped somefunction we expect a good error message
 """
         FSharp fsharpSource
         |> asExe